fix tests for package rename
